rename reporting snapshot record router
